+++ packages/eslint-plugin-gestalt/src/__fixtures__/no-workflow-status-icon/valid/no-match-color.js
@@ -1,0 +1,7 @@
+import { Icon } from 'gestalt';
+
+export default function TestElement() {
+  return (
+    <Icon icon="workflow-status-problem" color="dark" />
+  );
+}