+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 62.5.0 (Aug 1, 2022)
+
+### Minor
+
+- Eslint plugin: Avoid "workflow-status-[..]" on Icon (#2228)
+
+## 62.4.8 (Aug 1, 2022)
+
+### Patch
+
+- Internal: replace flow-lint suppressions with FlowFixMe (#2241)
+
 ## 62.4.7 (Aug 1, 2022)
 
 ### Patch